# Allow a body for GET requests