--- a/header_regression_test.go
+++ b/header_regression_test.go
@@ -81,11 +81,7 @@ func issue6VerifyRequestHeader(t *testing.T, h *RequestHeader, contentType strin
 	if string(h.Method()) != method {
 		t.Fatalf("unexpected method: %q. Expecting %q", h.Method(), method)
 	}
-	if method != MethodGet {
-		if h.ContentLength() != contentLength {
-			t.Fatalf("unexpected content-length: %d. Expecting %d. method=%q", h.ContentLength(), contentLength, method)
-		}
-	} else if h.ContentLength() != 0 {
-		t.Fatalf("unexpected content-length for GET method: %d. Expecting 0", h.ContentLength())
+	if h.ContentLength() != contentLength {
+		t.Fatalf("unexpected content-length: %d. Expecting %d. method=%q", h.ContentLength(), contentLength, method)
 	}
 }